made it so message is not deleted

+++ commands/mod/warn.js
@@ -223,9 +223,13 @@ async function warn(message, args, client) {
 
       if (message.channel.parent.id === "709806849725038634") {
       } else {
-        message.delete().catch((error) => {
-          console.log(error);
-        });
+        if(args[1]==="how dare you ping me, just spam my dm's or something smhh xD")
+        {}
+        else{
+          message.delete().catch((error) => {
+            console.log(error);
+          });
+        }
       }
     }
   });
